+++ sdk-libs/program-test/src/test_batch_forester.rs
@@ -169,7 +169,7 @@ pub async fn create_append_batch_ix_data<Rpc: RpcConnection>(
 
         // TODO: remove unwraps
         let circuit_inputs =
-            get_batch_append_with_proofs_inputs::<{ DEFAULT_BATCH_STATE_TREE_HEIGHT }>(
+            get_batch_append_with_proofs_inputs::<{ DEFAULT_BATCH_STATE_TREE_HEIGHT as usize }>(
                 old_root,
                 merkle_tree_next_index as u32,
                 batch_update_leaves,
@@ -328,7 +328,7 @@ pub async fn get_batched_nullify_ix_data<Rpc: RpcConnection>(
     // local_leaves_hashchain is only used for a test assertion.
     let local_nullifier_hashchain = create_hash_chain_from_slice(&nullifiers).unwrap();
     assert_eq!(leaves_hashchain, local_nullifier_hashchain);
-    let inputs = get_batch_update_inputs::<{ DEFAULT_BATCH_STATE_TREE_HEIGHT }>(
+    let inputs = get_batch_update_inputs::<{ DEFAULT_BATCH_STATE_TREE_HEIGHT as usize }>(
         old_root,
         tx_hashes,
         leaves.to_vec(),
@@ -861,25 +861,26 @@ pub async fn create_batch_update_address_tree_instruction_data_with_proof<
 
         low_element_proofs.push(non_inclusion_proof.low_address_proof.to_vec());
     }
-    let inputs = get_batch_address_append_circuit_inputs::<{ DEFAULT_BATCH_ADDRESS_TREE_HEIGHT }>(
-        start_index,
-        current_root,
-        low_element_values,
-        low_element_next_values,
-        low_element_indices,
-        low_element_next_indices,
-        low_element_proofs,
-        addresses,
-        indexer
-            .get_subtrees(merkle_tree_pubkey.to_bytes())
-            .unwrap()
-            .try_into()
-            .unwrap(),
-        leaves_hashchain,
-        batch_start_index,
-        batch.zkp_batch_size as usize,
-    )
-    .unwrap();
+    let inputs =
+        get_batch_address_append_circuit_inputs::<{ DEFAULT_BATCH_ADDRESS_TREE_HEIGHT as usize }>(
+            start_index,
+            current_root,
+            low_element_values,
+            low_element_next_values,
+            low_element_indices,
+            low_element_next_indices,
+            low_element_proofs,
+            addresses,
+            indexer
+                .get_subtrees(merkle_tree_pubkey.to_bytes())
+                .unwrap()
+                .try_into()
+                .unwrap(),
+            leaves_hashchain,
+            batch_start_index,
+            batch.zkp_batch_size as usize,
+        )
+        .unwrap();
     let client = Client::new();
     let circuit_inputs_new_root = bigint_to_be_bytes_array::<32>(&inputs.new_root).unwrap();
     let inputs = to_json(&inputs);
